refactor(runtime): lift the connection-gate core out of the coverage-excluded shell (#751)
The hook-transport gate tests pinned a hand-copied mirror of reducer_task's
gate + reconcile (`drive_hook_gate`), not the production path — and reducer_task
is excluded from codecov AND mutants by file path, so a drift in the real gate
predicate or reconcile placement would red nothing. This is the exact
false-green class #741/#751 flagged.

Extract `event_source` + `apply_gated_event` + `reconcile_sweep_tick` into a new,
NON-excluded `runtime::gate` module that BOTH `reducer_task` and the tests call.
The async shell stays thin (sans-IO — time is a parameter); the gate decision is
now covered and mutation-tested, so a drift reds a test instead of slipping
through. Follows the functional-core/imperative-shell + sans-IO practice the repo
already applies to `event_source` and `headless_loop`'s injected ctrl_c seam.

No behavior change: the three hook tests + the event_source test move verbatim
and still pass, now driving the REAL functions rather than a copy.

Closes #751.
